Add live publish-boundary audit for the research pages
The unit tests prove the cutoff FUNCTION is correct against injected dates.
They cannot prove the DEPLOYED pages honour it: the sessions on a real page
come from a real database through a real cache, and the failure being
guarded against — a live session on a public URL — is a property of the
rendered HTML. A green unit suite and a leaking page are compatible states.

So this reads what the server actually served and asserts the boundary on
the page itself: every date in the article body must precede today in ET.
It also checks that no rate is stated without its denominator, that no
unrounded float reaches a reader, and that the historical disclaimer is
present.

Read-only and UNAUTHENTICATED — these pages are public, so it needs no
Clerk user and no cleanup, and it exercises exactly the surface an
anonymous crawler sees rather than an authenticated approximation.

Three traps encoded:

  - Dates are parsed in BOTH forms. The table prints "August 19, 2026" and
    the JSON-LD prints "2026-08-19"; understanding only one would pass a
    page leaking through the other.
  - Numeric checks run against the <article> with scripts stripped.
    Scanning whole-page HTML sweeps up Next.js's inline bundle and flight
    payload, which are full of long floats and date-shaped strings — a gate
    that cries wolf is a gate that gets ignored.
  - A 200 that is not the research page reports HARNESS, never a product
    verdict, and zero parsed dates is HARNESS too: "the check did not run"
    must never read as "the check passed".

A 404 under the coverage floor is reported as INFO rather than FAIL — that
is the designed behaviour for a thin rail — but it IS reported, so "half
the sitemap 404s" cannot hide behind a green run.

Pre-deploy baseline run against production: hub 404, exits 1. Correct — the
pages are not deployed yet, and the gate fails closed.

+++ src/lib/research/gamma-levels.ts
@@ -100,7 +100,9 @@ async function loadWindowSamples(
           walls,
           gamma_flip,
           bucket_time,
-          NTILE($3) OVER (PARTITION BY session_ymd ORDER BY bucket_time) AS slot
+          -- Explicit ::int — a bind parameter reaches Postgres untyped, and NTILE will not take
+          -- the text it would otherwise be inferred as.
+          NTILE($3::int) OVER (PARTITION BY session_ymd ORDER BY bucket_time) AS slot
         FROM vector_wall_history
         WHERE ticker = $1 AND session_ymd = ANY($2::date[])
       ) t
